feat(network-manager): expose clientId and transportId conversion methods
ConnectionManager contains a useful map of clientId to transportId and vice versa. Exposing these methods can be useful for a developer to have access to. An example is, when using Facepunch.Steamworks and the community transport, you can use the clientId to get the steamId of that client, which can then be used as a unique identifier.